docs: add production readiness audit and roadmap
- Update ANALYSIS.md with full-stack security, reliability, and operational analysis

- Update PRODUCTIONREADY.md with 5.38/10 score and NO-GO verdict

- Update ROADMAP.md with prioritized P0-P3 milestones

- Include prior accumulated code changes across internal/

+++ internal/raft/multiregion.go
@@ -327,7 +327,7 @@ func (m *MultiRegionManager) UpdateReplicationStatus(regionID string, matchIndex
 
 	// Calculate lag
 	if m.node != nil {
-		status.ReplicationLag = time.Duration(m.node.lastLogIndex()-matchIndex) * time.Millisecond
+		status.ReplicationLag = time.Duration(int64(m.node.lastLogIndex()-matchIndex)) * time.Millisecond // #nosec G115 -- log index delta is bounded by Raft cluster size.
 	}
 
 	status.LastContact = time.Now()
